🎨 Split into files per package manager
